✨ feat(src): Update MRIden.py: add RAG

+++ src/MRIden.py
@@ -2,8 +2,8 @@
 Author       : Radon
 Date         : 2025-02-12 20:23:29
 LastEditors  : Radon
-LastEditTime : 2025-07-28 17:13:56
-Description  : 提示两个LLM进行MR识别和校对
+LastEditTime : 2025-07-30 16:04:47
+Description  : Prompt iden llm & cali llm to identify and calibrate metamorphic relation.
 """
 
 import argparse
@@ -12,7 +12,7 @@ import os
 import shutil
 
 from langchain_chroma import Chroma
-from langchain_openai import OpenAIEmbeddings
+from langchain_google_genai import GoogleGenerativeAIEmbeddings
 
 from utils import *
 from wrappers.anthropic import Anthropic
@@ -21,23 +21,20 @@ from wrappers.openai import OpenAI
 
 
 def check_config(args: argparse.Namespace):
-    """检查配置文件是否合法
+    """Check the validity of the configuration file and prepare the output directory.
 
     Parameters
     ----------
     args : argparse.Namespace
-        命令行参数集
+        Command line arguments containing the path of the configuration file.
     """
-    # 检查配置文件是否存在
     if not os.path.exists(args.config):
         FATAL(f"File not found: {args.config}")
-
-    # 读取配置文件
     config = dict()
     with open(args.config, "r") as f:
         config = json.load(f)
 
-    # 如果config中没有temperature字段, 使用默认值0.5
+    # If the key temperature is not exist, use default value 0.5
     if "temperature" not in config["identifier"].keys():
         WARNF('Key "temperature" not found in config file for "identifier", using default value: 0.5')
         config["identifier"]["temperature"] = 0.5
@@ -45,7 +42,7 @@ def check_config(args: argparse.Namespace):
         WARNF('Key "temperature" not found in config file for "calibrator", using default value: 0.5')
         config["calibrator"]["temperature"] = 0.5
 
-    # 如果config中没有stream字段, 使用默认值False
+    # If the key stream is not exist, use default value False
     if "stream" not in config["identifier"].keys():
         WARNF('Key "stream" not found in config file for "identifier", using default value: False')
         config["identifier"]["stream"] = False
@@ -53,33 +50,31 @@ def check_config(args: argparse.Namespace):
         WARNF('Key "stream" not found in config file for "calibrator", using default value: False')
         config["calibrator"]["stream"] = False
 
-    # 检查输出目录是否存在, 如果存在则报错, 提示用户需要先删掉该目录
+    # Prepare the output directory
     out_dir = config["output"]
-    shutil.rmtree(out_dir, ignore_errors=True)  # NOTE: Just for testing ...
-    if os.path.exists(out_dir):
+    if args.remove_exist_outdir:
+        shutil.rmtree(out_dir, ignore_errors=True)
+    elif os.path.exists(out_dir):
         FATAL(f"Output directory already exists: {out_dir}, please remove it first.")
-    os.makedirs(out_dir)
-
-    # 将用户的输入配置文件复制到输出目录下
-    shutil.copy(args.config, os.path.join(out_dir, "config.json"))
+    os.makedirs(out_dir, exist_ok=True)
 
 
 def setup_openai(config: dict, role: str) -> OpenAI:
-    """初始化以OpenAI为框架的聊天对象, 主要是GPT, DeepSeek等系列模型
+    """Initialize an OpenAI chat object
 
     Parameters
     ----------
     config : dict
-        用户输入的配置信息
+        Configuration provided by user
     role : str
-        GPT模型的角色, 可以是identifier或calibrator
+        Role of llm, which is identifier or calibrator
 
     Returns
     -------
     OpenAI
-        封装的OpenAI聊天对象
+        OpenAI chat object
     """
-    # 初始化以OpenAI为框架的聊天对象
+    # Initialize OpenAI chat object
     openai_obj = OpenAI(
         base_url=config[role]["base_url"],
         api_key=config[role]["api_key"],
@@ -88,10 +83,10 @@ def setup_openai(config: dict, role: str) -> OpenAI:
         stream=config[role]["stream"],
     )
 
-    driver_name = config["driver_name"]  # 待测驱动程序名称
-    spec = read_file(config["specification"])  # 规约说明文件的内容
+    driver_name = config["driver_name"]
+    spec = read_file(config["specification"])   # Specification content
 
-    # 设置模型的系统提示信息
+    # Set system prompt
     fn = config[role]["prompts"]["system"]
     sys_prompt = str()
     with open(fn, "r", encoding="utf-8") as f:
@@ -104,21 +99,21 @@ def setup_openai(config: dict, role: str) -> OpenAI:
 
 
 def setup_anthropic(config: dict, role: str) -> Anthropic:
-    """初始化以Anthropic为框架的聊天对象, 主要是Claude系列模型
+    """Initialize an Anthropic chat object
 
     Parameters
     ----------
     config : dict
-        用户输入的配置信息
+        Configuration provided by user
     role : str
-        以Anthropic为框架的角色, 可以是identifier或calibrator
+        Role of llm, which is identifier or calibrator
 
     Returns
     -------
     Anthropic
-        封装的Anthropic聊天对象
+        Anthropic chat object
     """
-    # 初始化以Anthropic为框架的聊天对象
+    # Initialize Anthropic chat object
     anthropic_obj = Anthropic(
         base_url=config[role]["base_url"],
         api_key=config[role]["api_key"],
@@ -127,10 +122,10 @@ def setup_anthropic(config: dict, role: str) -> Anthropic:
         stream=config[role]["stream"],
     )
 
-    driver_name = config["driver_name"]  # 待测驱动程序名称
-    spec = read_file(config["specification"])  # 规约说明文件的内容
+    driver_name = config["driver_name"]
+    spec = read_file(config["specification"])   # Specification content
 
-    # 设置模型的系统提示信息
+    # Set system prompt
     fn = config[role]["prompts"]["system"]
     sys_prompt = str()
     with open(fn, "r", encoding="utf-8") as f:
@@ -143,22 +138,21 @@ def setup_anthropic(config: dict, role: str) -> Anthropic:
 
 
 def setup_googleai(config: dict, role: str) -> GoogleAI:
-    """初始化以GoogleAI为框架的聊天对象, 主要是调用Gemini等模型
-    官方库名为google.genai
+    """Initialize a GoogleAI chat object
 
     Parameters
     ----------
     config : dict
-        用户输入的配置信息
+        Configuration provided by user
     role : str
-        对象的角色, 可以是identifier或calibrator
+        Role of llm, which is identifier or calibrator
 
     Returns
     -------
     GoogleAI
-        封装的以GoogleAI为框架的聊天对象
+        GoogleAI chat object
     """
-    # 初始化Gemini聊天模型
+    # Initialize googleai chat model
     googleai_obj = GoogleAI(
         base_url=config[role]["base_url"],
         api_key=config[role]["api_key"],
@@ -167,10 +161,10 @@ def setup_googleai(config: dict, role: str) -> GoogleAI:
         stream=config[role]["stream"],
     )
 
-    driver_name = config["driver_name"]  # 待测驱动程序名称
-    spec = read_file(config["specification"])  # 规约说明文件的内容
+    driver_name = config["driver_name"]
+    spec = read_file(config["specification"])  # Specification content
 
-    # 设置模型的系统提示信息
+    # Set system prompt
     fn = config[role]["prompts"]["system"]
     sys_prompt = str()
     with open(fn, "r", encoding="utf-8") as f:
@@ -199,26 +193,31 @@ def loop(
     config : dict
         Configuration provided by user
     """
-    iden_prompts = list()  # identifier提示词列表
-    cali_prompts = list()  # calibrator提示词列表
-    idx_iden_prompt = 0  # identifier提示词索引
-    idx_cali_prompt = 0  # calibrator提示词索引
-    spec = read_file(config["specification"])  # 读取规格说明文件的内容
-    dirver_name = config["driver_name"]  # 驱动程序的名称
-    prev_mr = str()  # 上一次的蜕变关系内容
-    mr = str()  # 当前的蜕变关系候选内容
-    iterations = 0  # 讨论轮数
-    gen_success = False  # 是否成功生成MR
-
-    # 读取identifier和calibrator的提示词
+    # Add user prompts to the list
+    iden_prompts = list()  # Prompt list for identifier
+    cali_prompts = list()  # Prompt list for calibrator
     for fn in config["identifier"]["prompts"]["user"]:
-        with open(fn, "r") as f:
-            iden_prompts.append(f.read())
+        iden_prompts.append(Path(fn).read_text(encoding="utf-8"))
     for fn in config["calibrator"]["prompts"]["user"]:
-        with open(fn, "r") as f:
-            cali_prompts.append(f.read())
+        cali_prompts.append(Path(fn).read_text(encoding="utf-8"))
+
+    # Retrieve relevant documents from the external corpus if exists
+    spec = Path(config["specification"]).read_text(encoding="utf-8")
+    retrieved_prompt = str()
+    if vector_db is not None:
+        retrieved_docs = vector_db.similarity_search(query=spec)
+        docs_content = "\n\n---\n\n".join(doc.page_content for doc in retrieved_docs)
+        if len(docs_content) > 0:
+            retrieved_prompt = f"You can also refer to the following documents:\n\n{docs_content}"
 
     # Prompt identifier and calibrator to identify and calibrate metamorphic relations iteratively
+    prev_mr = str()  # Previous metamorphic relation content
+    mr = str()  # Latest metamorphic relation
+    gen_success = False  # Whether the MR generation is successful
+    idx_iden_prompt = 0  # identifier prompt index
+    idx_cali_prompt = 0  # calibrator prompt index
+    iterations = 0
+    dirver_name = config["driver_name"]  # Driver name
     while iterations < config["max_iter"]:
         ACTF(f"Iterations: {iterations + 1}")
         ACTF(f"Asking identifier ({config['identifier']['model']}) ...")
@@ -227,18 +226,15 @@ def loop(
         iden_prompt = iden_prompt.replace("[MR generated by calibrator]", mr)
         iden_prompt = iden_prompt.replace("[Driver name]", dirver_name)
 
-        # Retrieve relevant documents from the external corpus if exists
-        if vector_db is not None:
-            retrieved_docs = vector_db.similarity_search(query=spec)
-            docs_content = "\n\n".join(doc.page_content for doc in retrieved_docs)
-            if len(docs_content) > 0:
-                retrieved_prompt = f"You can also refer to the following documents:\n\n{docs_content}"
-                iden_prompt = f"{iden_prompt}\n\n{retrieved_prompt}"
+        # Add retrieved documents to the identifier prompt if exists
+        if len(retrieved_prompt) > 0:
+            iden_prompt += "\n\n" + retrieved_prompt
 
         # Prompt identifier to generate an metamorphic relation
         iden_response = identifier.chat(iden_prompt)
 
-        # 如果输出的内容中不存在代码块, 认为identifier没有继续改进MR, 跳出循环
+        # If identifier's output does not contain a code block, or not further improved
+        # MR, break the loop
         prev_mr = mr
         mr = get_first_code_block(iden_response, {"markdown", "md"})
         if len(mr) == 0:
@@ -251,7 +247,8 @@ def loop(
         cali_prompt = cali_prompt.replace("[Driver name]", dirver_name)
         cali_response = calibrator.chat(cali_prompt)
 
-        # 如果calibrator的输出内容中不存在代码块, 或输出的是"correct", 视作calibrator认为此次的MR正确反应了待测对象的属性, 跳出循环
+        # If calibrator's output does not contain a code block, or its output is "correct",
+        # then we consider the MR is successfully generated
         prev_mr = mr
         mr = get_first_code_block(cali_response, {"markdown", "md"})
         if len(mr) == 0 or cali_response.lower() == "correct":
@@ -259,16 +256,17 @@ def loop(
             break
         OKF("Got the MR generated by calibrator!")
 
-        # 更新identifier和calibrator的提示词索引, 如果索引超出范围, 则不再更新
+        # Update prompt index of iden llm & cali llm for the next iteration
         if idx_iden_prompt < len(config["identifier"]["prompts"]["user"]) - 1:
             idx_iden_prompt += 1
         if idx_cali_prompt < len(config["calibrator"]["prompts"]["user"]) - 1:
             idx_cali_prompt += 1
 
-        # 更新迭代轮数计数
+        # Update iteration count
         iterations += 1
 
-    # 将和identifier及calibrator的对话记录保存至markdown和json文件, 并将两个LLM的最终讨论结果保存至output下的mr_final.md
+    # Save chat messages of iden llm & cali llm to the output directory
+    # Save the final MR to the output directory
     identifier.save_messages(os.path.join(config["output"], "iden_messages.md"))
     identifier.save_messages(os.path.join(config["output"], "iden_messages.json"))
     calibrator.save_messages(os.path.join(config["output"], "cali_messages.md"))
@@ -319,8 +317,8 @@ def main(config: dict):
     # Load the external corpus if exists
     vector_db = None
     if "corpus" in config.keys():
-        chroma_dir = os.path.join(config["corpus"]["path"], "chroma")
-        embeddings = OpenAIEmbeddings(model=config["corpus"]["embeddings"])
+        chroma_dir = os.path.join(config["corpus"], "chroma")
+        embeddings = GoogleGenerativeAIEmbeddings(model="models/gemini-embedding-001")
         vector_db = Chroma(persist_directory=chroma_dir, embedding_function=embeddings)
 
     # Prompt iden llm & cali llm to identify and calibrate metamorphic relations
@@ -331,14 +329,15 @@ def main(config: dict):
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
     parser.add_argument("--config", type=str, required=True, help="Path of configuration json file.")
+    parser.add_argument("--remove-exist-outdir", action="store_true", help="Remove the output directory if it exists.")
     args = parser.parse_args()
 
-    # 检查命令行参数是否合法
+    # Check validity of arguments
     ACTF("Checking arguments...")
     check_config(args)
     OKF("Arguments are valid.")
 
-    # 读取配置文件
+    # Read the configuration file
     config = dict()
     with open(args.config, "r") as f:
         config = json.load(f)
